Refactor AdapativeThreshold transform to custom_transforms file
